backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.30
sqlmodel==0.0.16
pydantic==2.7.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
asyncpg==0.29.0
PyJWT==2.9.0
aiosqlite==0.20.0
